+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "structured-output-test"
+name = "structured-rag"
 version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
@@ -7,7 +7,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
-dspy_ai = { path = "/Users/cshorten/Desktop/dspy/dist/dspy_ai-2.4.10-py3-none-any.whl" }
+dspy_ai = "^2.4.13"
 ollama = "^0.2.1"
 matplotlib = "^3.9.1"
