+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="structured-rag",
+    version="0.1",
+    packages=find_packages(where="src"),
+    package_dir={"": "src"},
+)
